kcal int to double

+++ src/main/java/org/example/devops_mini_back/dto/ExerciseFavorite/ExerciseFavoriteResponseDto.java
@@ -12,6 +12,6 @@ public class ExerciseFavoriteResponseDto {
     private int userId;
     private String exerciseName;
     private String youtubeId;
-    private int kcal;
+    private double kcal;
     private String picture;
 }
